apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
google-cloud-firestore==2.14.0
firebase-admin==6.4.0
stripe==8.1.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-json-logger==2.0.7
httpx==0.26.0
